requirements.txt
docx
python-docx
google-generativeai